tools/pyproject.toml
[tool.poetry]
name = "cgdevxcli"
version = "0.3.0"
description = ""
authors = ["Alexander Ulyanov <alexu@cloudgeometry.io>", "Mikhail Turetskii <mturetsky@cloudgeometry.io>"]
readme = "README.md"
packages = [{ include = "cli" }]

[tool.poetry.scripts]
cgdevxcli = "cli.__main__:entry_point"
build = "cli.build:install"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
click = "8.1.7"
cloup = "3.0.2"
PyInstaller = "5.13.2"
awscli = "1.29.56"
boto3 = "1.28.56"
configparser = "6.0.0"
pyyaml = "6.0.1"
kubernetes = "27.2.0"
azure-identity = "1.14.0"
azure-storage-blob = "12.17.0"
azure-mgmt-resource = "23.0.1"
azure-mgmt-authorization = "^4.0.0"
azure-mgmt-dns = "^8.1.0"
azure-mgmt-privatedns = "^1.1.0"
azure-mgmt-containerservice = "^26.0.0"
azure-mgmt-storage = "^21.1.0"
azure-mgmt-subscription = "^3.1.1"
azure-mgmt-compute = "^30.5.0"
gcloud = "0.18.3"
gitpython = "3.1.32"
ghrepo = "0.7.0"
dnspython = "2.4.2"
hvac = "1.2.1"
alive-progress = "^3.1.5"
kr8s = "^0.14.2"
google-cloud-storage = "^2.15.0"
google-cloud-dns = "^0.35.0"
google-auth = "^2.28.2"
google-api-python-client = "^2.121.0"

[tool.poetry.dev-dependencies]
flake8 = "6.1.0"
flake8-class-attributes-order = "0.1.3"
flake8-docstrings = "1.7.0"
pydocstyle = "6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
